apps/physics-service/pyproject.toml
[project]
name = "physics-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "dimbo1324", email = "dimaprihodko180@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "pydantic-settings>=2.12.0",
]
[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"
[dependency-groups]
dev = [
    "grpcio-tools>=1.76.0",
    "hypothesis>=6.151.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-html>=4.2.0",
]
[tool.uv]
package = false